update retry proposal
Currently, the retry proposal specifies sending "grpc-previous-rpc-attempts"
metadata on each RPC beyond the first RPC.

However, this makes it hard to use: the absence of the metadata could either be
RPC attempt #0, or an RPC (of any attempt #) using a version of the gRPC library
before this logic is implemented. (or, the version of the client library that
was updated to use this field)

That is: we could never tell, for example, %QPS that are related to retries
because we don't know what our sample size is. By specifying first RPC must send
grpc-previous-rpc-attempts=0, we can do #rpcs-whose-attempts>0 / #rpcs-with-any-attempts.