Revamp waiting-for-xds-config
Waiting for xds config is not just a one-shot affair at server startup,
since xds can delete resources after started. That caused a pretty
strong rework to the behavior, including ditching
listen()-but-don't-accept(). We still need to validate that
accept()+close() won't cause poor behavior for clients such as Envoy,
but the options available are limited.